+++ src/Filament/Resources/PopupFollowUpFlowResource.php
@@ -2,28 +2,28 @@
 
 namespace Dashed\DashedPopups\Filament\Resources;
 
-use BackedEnum;
 use UnitEnum;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteAction;
-use Filament\Actions\DeleteBulkAction;
+use BackedEnum;
+use Filament\Tables\Table;
+use Filament\Schemas\Schema;
 use Filament\Actions\EditAction;
+use Filament\Resources\Resource;
+use Filament\Actions\DeleteAction;
+use Filament\Actions\BulkActionGroup;
+use Filament\Forms\Components\Toggle;
+use Filament\Actions\DeleteBulkAction;
 use Filament\Forms\Components\Builder;
 use Filament\Forms\Components\Repeater;
-use Filament\Forms\Components\RichEditor;
 use Filament\Forms\Components\Textarea;
-use Filament\Forms\Components\TextInput;
-use Filament\Forms\Components\Toggle;
-use Filament\Resources\Resource;
-use Filament\Schemas\Components\Section;
-use Filament\Schemas\Schema;
 use Filament\Tables\Columns\IconColumn;
 use Filament\Tables\Columns\TextColumn;
-use Filament\Tables\Table;
+use Filament\Forms\Components\TextInput;
+use Filament\Schemas\Components\Section;
+use Filament\Forms\Components\RichEditor;
 use Dashed\DashedPopups\Models\PopupFollowUpFlow;
-use Dashed\DashedPopups\Filament\Resources\PopupFollowUpFlowResource\Pages\CreatePopupFollowUpFlow;
 use Dashed\DashedPopups\Filament\Resources\PopupFollowUpFlowResource\Pages\EditPopupFollowUpFlow;
 use Dashed\DashedPopups\Filament\Resources\PopupFollowUpFlowResource\Pages\ListPopupFollowUpFlows;
+use Dashed\DashedPopups\Filament\Resources\PopupFollowUpFlowResource\Pages\CreatePopupFollowUpFlow;
 
 class PopupFollowUpFlowResource extends Resource
 {